Split chart from index.js to start modularizing before onboarding notification
