+++ public/index.html
@@ -22,14 +22,14 @@
 <body>
   <div class="wrapper">
     <div class="total">
-      <div class="total">Your total is: $<span id="total">0</span></div>
+      <div class="total">Balance: $<span id="total">0</span></div>
     </div>
 
     <div class="form">
       <input type="text" id="t-name" placeholder="Name of transaction" />
       <input type="number" min="0" id="t-amount" placeholder="Transaction amount" />
-      <button id="add-btn"><i class="fa fa-plus buttons"></i> Add Funds</button>
-      <button id="sub-btn"><i class="fa fa-minus"></i> Subtract Funds</button>
+      <button id="add-btn"><i class="fa fa-plus buttons"></i>Deposit</button>
+      <button id="sub-btn"><i class="fa fa-minus"></i>Withdraw</button>
       <p class="error"></p>
     </div>
 
@@ -38,6 +38,7 @@
         <thead>
           <th>Transaction</th>
           <th>Amount</th>
+          <th>Date-Time</th>
         </thead>
         <tbody id="tbody">
           
@@ -49,7 +50,8 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
-  <script src="js/index.js"></script>
+  <script type="module" src="js/chart.js"></script>
+  <script type="module" src="js/index.js"></script>
   <script>
     (async function() {
       if("serviceWorker" in navigator) { // see if serviceWorker exists on the global navigator object.
