+++ public/js/index.js
@@ -1,5 +1,10 @@
-let transactions = [];
-let myChart;
+
+export var transactions = [];
+// let myChart;
+
+import {populateChart } from './chart.js';
+
+
 
 fetch("/api/transaction")
   .then(response => {
@@ -34,49 +39,50 @@ function populateTable() {
     tr.innerHTML = `
       <td>${transaction.name}</td>
       <td>${transaction.value}</td>
+      <td>${transaction.date}</td>
     `;
 
     tbody.appendChild(tr);
   });
 }
 
-function populateChart() {
-  // copy array and reverse it
-  let reversed = transactions.slice().reverse();
-  let sum = 0;
+// function populateChart() {
+//   // copy array and reverse it
+//   let reversed = transactions.slice().reverse();
+//   let sum = 0;
 
-  // create date labels for chart
-  let labels = reversed.map(t => {
-    let date = new Date(t.date);
-    return `${date.getMonth() + 1}/${date.getDate()}/${date.getFullYear()}`;
-  });
+//   // create date labels for chart
+//   let labels = reversed.map(t => {
+//     let date = new Date(t.date);
+//     return `${date.getMonth() + 1}/${date.getDate()}/${date.getFullYear()}`;
+//   });
 
-  // create incremental values for chart
-  let data = reversed.map(t => {
-    sum += parseInt(t.value);
-    return sum;
-  });
+//   // create incremental values for chart
+//   let data = reversed.map(t => {
+//     sum += parseInt(t.value);
+//     return sum;
+//   });
 
-  // remove old chart if it exists
-  if (myChart) {
-    myChart.destroy();
-  }
+//   // remove old chart if it exists
+//   if (myChart) {
+//     myChart.destroy();
+//   }
 
-  let ctx = document.getElementById("myChart").getContext("2d");
+//   let ctx = document.getElementById("myChart").getContext("2d");
 
-  myChart = new Chart(ctx, {
-    type: 'line',
-      data: {
-        labels,
-        datasets: [{
-            label: "Total Over Time",
-            fill: true,
-            backgroundColor: "#6666ff",
-            data
-        }]
-    }
-  });
-}
+//   myChart = new Chart(ctx, {
+//     type: 'line',
+//       data: {
+//         labels,
+//         datasets: [{
+//             label: "Balance over Time",
+//             fill: true,
+//             backgroundColor: "#6666ff",
+//             data
+//         }]
+//     }
+//   });
+// }
 
 function sendTransaction(isAdding) {
   let nameEl = document.querySelector("#t-name");
